refactor : player와 resultItem의 중복을 새로운 클래스로 분리합니다
이름의 유효성 검사 기능을 LadderItem 클래스를 만들어 담당하게 하고,
player와 resultItem에서 상속받도록 하였습니다.
