+++ src/main/java/ladder/domain/Player.java
@@ -2,13 +2,10 @@ package ladder.domain;
 
 import java.util.*;
 
-public class Player {
-    private static final int MAXIMUM_NAME_LENGTH = 5;
-    private static final int MINIMUM_NANE_LENGTH = 1;
+public class Player extends LadderItem {
     private static final String FORBIDDEN_NAME = "all";
     private static final int DEFAULT_POSITION = 0;
 
-    private final String name;
     private int position;
 
     Player(final String name) {
@@ -16,22 +13,9 @@ public class Player {
     }
 
     Player(String name, int position) {
-        nullCheck(name);
-        this.name = getTrimmed(name);
-        this.position = position;
-    }
-
-    private void nullCheck(String name) {
-        if (name == null) {
-            throw new IllegalArgumentException("이름에는 null이 입력될 수 없습니다.");
-        }
-    }
-
-    private String getTrimmed(String name) {
-        name = name.trim();
+        super(name);
         validateCommandName(name);
-        validateNameLength(name);
-        return name;
+        this.position = position;
     }
 
     private void validateCommandName(String name) {
@@ -40,34 +24,24 @@ public class Player {
         }
     }
 
-    private void validateNameLength(String name) {
-        if ((name.length() > MAXIMUM_NAME_LENGTH) || (name.length() < MINIMUM_NANE_LENGTH)) {
-            throw new IllegalArgumentException("이름의 길이는 1자 이상, 5자 이하여야 합니다.");
-        }
-    }
-
     Map<String, ResultItem> stepDown(Ladder ladder) {
         Map<String, ResultItem> ladderingResult = new LinkedHashMap<>();
 
-        ladderingResult.put(name, ladder.answerResult(position));
+        ladderingResult.put(getName(), ladder.answerResult(position));
         return ladderingResult;
-    }
-
-    @Override
-    public String toString() {
-        return this.name;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
         Player player = (Player) o;
-        return Objects.equals(name, player.name);
+        return position == player.position;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name);
+        return Objects.hash(super.hashCode(), position);
     }
 }
